Extract timing code in its crate
