+++ Cargo.toml
@@ -15,6 +15,7 @@ members = [
     "tezos/wrapper",
     "tezos/context",
     "tezos/new_context",
+    "tezos/timing",
     "networking",
     "shell",
     "storage",
