+++ rpc/Cargo.toml
@@ -26,8 +26,8 @@ shell = { path = "../shell" }
 storage = { path = "../storage" }
 tezos_api = { path = "../tezos/api" }
 tezos_context = { path = "../tezos/context" }
-tezos_new_context = { path = "../tezos/new_context" }
 tezos_messages = { path = "../tezos/messages" }
+tezos_timing = { path = "../tezos/timing" }
 tezos_wrapper = { path = "../tezos/wrapper" }
 rusqlite = "0.25.1"
