+++ tezos/timing/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "tezos_timing"
+version = "0.1.0"
+authors = ["Sebastien Chapuis <sebastien@chapu.is>"]
+edition = "2018"
+
+[dependencies]
+crypto = { path = "../../crypto" }
+serde = { version = "1.0", features = ["derive"] }
+once_cell = "1.7.2"
+crossbeam-channel = "0.5.1"
+rusqlite = { version = "0.25.1", features = [] }
